Rewritten HTTPServer to use size_t for data lengths.

+++ src/HTTPServer/HTTPConnection.h
@@ -51,7 +51,7 @@ public:
 	void Send(const cHTTPResponse & a_Response);
 	
 	/** Sends the data as the response (may be called multiple times) */
-	void Send(const void * a_Data, int a_Size);
+	void Send(const void * a_Data, size_t a_Size);
 
 	/** Sends the data as the response (may be called multiple times) */
 	void Send(const AString & a_Data) { Send(a_Data.data(), a_Data.size()); }
@@ -87,7 +87,7 @@ protected:
 
 	/** Number of bytes that remain to read for the complete body of the message to be received.
 	Valid only in wcsRecvBody */
-	int m_CurrentRequestBodyRemaining;
+	size_t m_CurrentRequestBodyRemaining;
 	
 	
 	// cSocketThreads::cCallback overrides:
